Big test rewrite, now includes test-utils
This addresses the two main issues I had with the tests:
1. I wasn't really testing whether optional/maybe generators were ever
omitted/nil, just that they fit the corresponding optional/maybe schema.
2. There was a substantial amount of boilerplate code to write
properties that were basically equivalent.

Number 1 is addressed by the `generator-sometimes-fits-schema?`
predicate, which makes it possible to check that an optional key is
sometimes missing or sometimes present, for example.

Number 2 is addressed by the `generator-fits-schema-prop`, which neatly
expresses one of the main requirements here.